FIX: position is updated after a character moves
When a character moved, its position attribute would not be
updated, thus triggering an error in the check_character_position
decorator. It called the move_to function a second time, returning
an error from the API.

Fixed by adding a function that updates the character position
whenever move_to is called.