REFACTOR: Add properties to characters and refactor character init
Handling cooldown was a problem if I started my program during their
cooldown period. It was also impossible to get the character's
real-time cooldown, because the cooldown duration provided by the API
is the last-known cooldown duration.
So even though enough time has passed and my character is not in
 cooldown, the data would say it is.

-Add cooldown_expiration and cooldown_duration to character properties
-Add some lines to calculate remaining colldown based on the date and
time on which the cooldown ends.
-Add an if-condition when creating an instance to check if characters
have any cooldown left