FIX: make_post_request returns a response, which is then checked for errors in higher functions
The function initially did not return the response if an error was
detected, because the error was tracked in the post request method.

Since each action can have different error status, let make post request
only handle exceptions or succesful responses. If the response contains
an error, the caller function will handle the returned response